+++ src/lang/en.ts
@@ -206,6 +206,9 @@ export const translation: Translation = {
             },
             description: 'Deploys to a ZIP file and sends it as attachment by mail via SMTP',
         },
+        nativescript: {
+            description: "Deploys a NativeScript project to a device or emulator",
+        },
         pipeline: {
             description: 'Pipes a list of sources files to a new destination, by using a script and sends the new file list to a target',
             noPipeFunction: "{0:trim,surround} implements no 'pipe()' function!",